requirements.txt
django-markdownx==4.0.7
django-zxcvbn-password==2.1.1
django-environ==0.11.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.1
psycopg[c]==3.1.18
